refactor: PictureTag 도메인 이동
