+++ src/main/java/com/onebyte/life4cut/album/controller/AlbumController.java
@@ -9,9 +9,9 @@ import com.onebyte.life4cut.album.controller.dto.UpdatePictureRequest;
 import com.onebyte.life4cut.auth.dto.CustomUserDetails;
 import com.onebyte.life4cut.common.web.ApiResponse;
 import com.onebyte.life4cut.common.web.EmptyResponse;
-import com.onebyte.life4cut.picture.domain.PictureTag;
 import com.onebyte.life4cut.picture.service.PictureService;
 import com.onebyte.life4cut.picture.service.dto.PictureDetailInSlot;
+import com.onebyte.life4cut.pictureTag.domain.PictureTag;
 import com.onebyte.life4cut.pictureTag.service.PictureTagService;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
