+++ src/main/java/com/onebyte/life4cut/pictureTag/repository/PictureTagRepository.java
@@ -1,7 +1,7 @@
 package com.onebyte.life4cut.pictureTag.repository;
 
-import com.onebyte.life4cut.picture.domain.PictureTag;
-import com.onebyte.life4cut.picture.domain.PictureTags;
+import com.onebyte.life4cut.pictureTag.domain.PictureTag;
+import com.onebyte.life4cut.pictureTag.domain.PictureTags;
 import java.util.List;
 
 public interface PictureTagRepository {
