+++ src/main/java/com/onebyte/life4cut/pictureTag/repository/PictureTagRepositoryImpl.java
@@ -1,10 +1,10 @@
 package com.onebyte.life4cut.pictureTag.repository;
 
-import static com.onebyte.life4cut.picture.domain.QPictureTag.pictureTag;
+import static com.onebyte.life4cut.pictureTag.domain.QPictureTag.pictureTag;
 import static org.springframework.util.StringUtils.hasText;
 
-import com.onebyte.life4cut.picture.domain.PictureTag;
-import com.onebyte.life4cut.picture.domain.PictureTags;
+import com.onebyte.life4cut.pictureTag.domain.PictureTag;
+import com.onebyte.life4cut.pictureTag.domain.PictureTags;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.annotation.Nonnull;
