+++ src/main/java/com/onebyte/life4cut/pictureTag/service/PictureTagService.java
@@ -2,7 +2,7 @@ package com.onebyte.life4cut.pictureTag.service;
 
 import com.onebyte.life4cut.album.exception.UserAlbumRolePermissionException;
 import com.onebyte.life4cut.album.repository.UserAlbumRepository;
-import com.onebyte.life4cut.picture.domain.PictureTag;
+import com.onebyte.life4cut.pictureTag.domain.PictureTag;
 import com.onebyte.life4cut.pictureTag.repository.PictureTagRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
