+++ src/test/java/com/onebyte/life4cut/album/controller/AlbumControllerTest.java
@@ -27,10 +27,10 @@ import com.onebyte.life4cut.album.controller.dto.UpdatePictureRequest;
 import com.onebyte.life4cut.common.annotation.WithCustomMockUser;
 import com.onebyte.life4cut.common.controller.ControllerTest;
 import com.onebyte.life4cut.fixture.PictureTagFixtureFactory;
-import com.onebyte.life4cut.picture.domain.vo.PictureTagName;
 import com.onebyte.life4cut.picture.repository.dto.PictureDetailResult;
 import com.onebyte.life4cut.picture.service.PictureService;
 import com.onebyte.life4cut.picture.service.dto.PictureDetailInSlot;
+import com.onebyte.life4cut.pictureTag.domain.vo.PictureTagName;
 import com.onebyte.life4cut.pictureTag.service.PictureTagService;
 import com.onebyte.life4cut.slot.domain.vo.SlotLayout;
 import com.onebyte.life4cut.slot.domain.vo.SlotLocation;
