+++ src/test/java/com/onebyte/life4cut/picture/domain/PictureTagTest.java
@@ -3,7 +3,8 @@ package com.onebyte.life4cut.picture.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.onebyte.life4cut.fixture.PictureTagFixtureFactory;
-import com.onebyte.life4cut.picture.domain.vo.PictureTagName;
+import com.onebyte.life4cut.pictureTag.domain.PictureTag;
+import com.onebyte.life4cut.pictureTag.domain.vo.PictureTagName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
