+++ src/test/java/com/onebyte/life4cut/picture/service/PictureServiceIntTest.java
@@ -21,14 +21,14 @@ import com.onebyte.life4cut.fixture.PictureTagRelationFixtureFactory;
 import com.onebyte.life4cut.fixture.SlotFixtureFactory;
 import com.onebyte.life4cut.fixture.UserAlbumFixtureFactory;
 import com.onebyte.life4cut.picture.domain.Picture;
-import com.onebyte.life4cut.picture.domain.PictureTag;
 import com.onebyte.life4cut.picture.domain.PictureTagRelation;
 import com.onebyte.life4cut.picture.domain.PictureTagRelations;
-import com.onebyte.life4cut.picture.domain.vo.PictureTagName;
 import com.onebyte.life4cut.picture.exception.PictureNotFoundException;
 import com.onebyte.life4cut.picture.repository.PictureRepositoryImpl;
 import com.onebyte.life4cut.picture.repository.PictureTagRelationRepositoryImpl;
 import com.onebyte.life4cut.picture.service.dto.PictureDetailInSlot;
+import com.onebyte.life4cut.pictureTag.domain.PictureTag;
+import com.onebyte.life4cut.pictureTag.domain.vo.PictureTagName;
 import com.onebyte.life4cut.pictureTag.repository.PictureTagRepositoryImpl;
 import com.onebyte.life4cut.slot.domain.Slot;
 import com.onebyte.life4cut.slot.repository.SlotRepositoryImpl;
